cttc-nr-fh-xr: (fixes #241) Replace non-existent uint type with size_t

+++ examples/cttc-nr-fh-xr.cc
@@ -2302,7 +2302,7 @@ ConfigureXrApp(NodeContainer& ueContainer,
     std::vector<Address> addresses;
     std::vector<InetSocketAddress> localAddresses;
 
-    for (uint j = 0; j < it->second.size(); j++)
+    for (size_t j = 0; j < it->second.size(); j++)
     {
         addresses.push_back(InetSocketAddress(address, port + j));
         // The sink will always listen to the specified ports
